+++ src/main/java/ua/com/agroswit/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UserService {
-    Optional<User> getById(Long id);
+    Optional<User> getById(Integer id);
     Optional<User> getByLogin(String name);
     List<User> getAll();
 }
